Use the main wiki configuration in case other value is not set at subwiki level #5 #17 * refactor * add script service to access collabora configurations
