+++ application-collabora-default/src/main/resources/META-INF/components.txt
@@ -5,3 +5,4 @@ com.xwiki.collabora.internal.DiscoveryManager
 com.xwiki.collabora.internal.configuration.CollaboraConfigurationSource
 com.xwiki.collabora.internal.configuration.MainCollaboraConfigurationSource
 com.xwiki.collabora.internal.configuration.DefaultCollaboraConfiguration
+com.xwiki.collabora.script.CollaboraConfigurationScriptService
